requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
requests
openai
